feat(router-store): add provider function for router store
