+++ modules/router-store/src/index.ts
@@ -45,3 +45,4 @@ export {
   MinimalRouterStateSerializer,
 } from './serializers/minimal_serializer';
 export { getSelectors, createRouterSelector } from './router_selectors';
+export { provideRouterStore } from './provide_router_store';
